server/requirements.txt
python-docx>=0.8.11
openpyxl>=3.1.2
reportlab>=4.0.4
python-pptx>=0.6.21
Pillow>=10.0.0
numpy>=1.24.0